add package createNews in tests
in this package splite CreateNewsTest class by logical
to some smaller classes
